requirements.txt
crewai==0.41.1
firecrawl-py
python-dotenv